Fix timing issue with GPhoto making it stuck in busy state after initial capture

+++ 3rdparty/indi-gphoto/gphoto_ccd.h
@@ -118,7 +118,7 @@ class GPhotoCCD : public INDI::CCD, public INDI::FocuserInterface
     void ShowExtendedOptions(void);
     void HideExtendedOptions(void);
 
-    float CalcTimeLeft();
+    double CalcTimeLeft();
     bool grabImage();
 
     char name[MAXINDIDEVICE];
